Refactor fixture cf. Colin's comment
