+++ components/tools/OmeroPy/test/integration/clitest/test_delete.py
@@ -288,15 +288,6 @@ class TestDelete(CLITest):
         assert self.query.find('Image', img.id.val)
 
     # Test combinations of include and exclude other than annotations
-    @pytest.fixture()
-    def simpleHierarchy(self):
-        proj = self.make_project()
-        dset = self.make_dataset()
-        img = self.update.saveAndReturnObject(self.new_image())
-        self.link(proj, dset)
-        self.link(dset, img)
-        return proj, dset, img
-
     def testExcludeNone(self, simpleHierarchy):
         proj, dset, img = simpleHierarchy
